moved frontpage and moviepage to folder pages
